feat: Implement athlete class enrollment and available classes screen
This commit introduces a complete feature for athletes to view and enroll in available classes. It includes a new dedicated screen, backend logic for enrollment, and updates to the view models and data models to support this functionality.

### Key Changes:

*   **New Feature: Available Classes Screen (`AvailableClassesScreen.kt`):**
    *   A new screen has been created to display all available classes to athletes.
    *   Athletes can filter classes by day of the week ("Todos", "Lun", "Mar", etc.).
    *   Each class card (`AvailableClassCard`) shows detailed information: name, time, coach, workout (if assigned), level, and available spots with a progress bar.
    *   Athletes can enroll ("Inscribirse") or cancel their enrollment ("Cancelar Inscripción") directly from the card.
    *   The UI provides clear states for loading, errors, and when no classes are available for a selected day.

*   **Enrollment Logic (Backend & Frontend):**
    *   **Use Cases:** Added `AvailableClassesUseCase`, `EnrollClassUseCase`, and `UnEnrollClassUseCase` to handle the business logic for fetching and managing enrollments.
    *   **ViewModel (`ClassViewModel`):** Updated to include new state flows (`availableClassesState`, `enrollmentState`) and functions (`loadAvailableClasses`, `enrollInClass`, `unenrollFromClass`) to drive the new screen.
    *   **Repository (`ClassRepositoryImpl`):** Implemented methods to call the new API endpoints: `availableClasses`, `enrollInClass`, and `unenrollInClass`.
    *   **API (`ClassApi.kt`):** Added suspend functions to communicate with the backend endpoints for `/api/classes/available` and `/api/classes/{classId}/enroll` (POST and DELETE).

*   **Data Models & UI State:**
    *   Introduced new data models to support the feature: `AvailableClassesResponse`, `EnrollmentResponse`, `UnEnrollmentResponse`, `ClassInfo`, and `Enrollment`.
    *   `ClassUiState` was expanded to include states for `AvailableClassesLoaded`, `EnrollmentLoading`, `EnrollmentSuccess`, and `UnenrollmentSuccess`.
    *   The `TodayClass` model now has a nullable `workout` to handle cases where a class for today does not yet have an assigned workout.

*   **Athlete Dashboard Enhancements:**
    *   The "Ver Clases Disponibles" button on the `AthleteDashboardScreen` now correctly navigates to the new `AvailableClasses` screen.
    *   The `TodayClassCard` has been updated to gracefully handle classes without an assigned workout, displaying a "Workout aún no asignado" message instead of crashing.
    *   Backend sorting for `upcomingClasses` in `AthleteRepository` was implemented to ensure classes are correctly ordered by day of the week and start time.

+++ shared/src/commonMain/kotlin/com/tepuytech/fitzon/domain/model/Workout.kt
@@ -6,6 +6,6 @@ data class Workout(
     val id: String,
     val title: String,
     val difficulty: String,
-    val duration: Long,
+    val duration: Int,
     val isCompleted: Boolean,
 )
